reworked on comments part 2
