+++ examples/analog_out/gen_voltage_wfm_int_clk.py
@@ -5,8 +5,6 @@ voltage samples to an Analog Output Channel using an internal
 sample clock.
 """
 
-from analog_out_helper import create_voltage_sample
-
 import nidaqmx
 from nidaqmx.constants import AcquisitionType
 
@@ -18,7 +16,7 @@ with nidaqmx.Task() as task:
         1000.0, sample_mode=AcquisitionType.FINITE, samps_per_chan=total_samples
     )
 
-    data = create_voltage_sample(5.0, total_samples)
+    data = [5.0 * i / total_samples for i in range(total_samples)]
     number_of_samples_written = task.write(data, auto_start=True)
     print(f"Generate {number_of_samples_written} voltage samples.")
     task.wait_until_done()
